refactor: seperate lecture
