+++ .babelrc
@@ -10,5 +10,8 @@
       }
     ]
   ],
-  "plugins": ["@emotion/babel-plugin"]
+  "plugins": [
+    ["styled-components", { "ssr": true, "displayName": true }],
+    "@emotion/babel-plugin"
+  ]
 }
